Split mutations into SubtitleEditor class
