+++ GuiSubtrans/Commands/BatchSubtitlesCommand.py
@@ -8,6 +8,7 @@ from PySubtrans.Options import Options
 from PySubtrans.SubtitleBatcher import SubtitleBatcher
 from PySubtrans.SubtitleProcessor import SubtitleProcessor
 from PySubtrans.SubtitleProject import SubtitleProject
+from PySubtrans.SubtitleEditor import SubtitleEditor
 
 import logging
 
@@ -30,20 +31,21 @@ class BatchSubtitlesCommand(Command):
         if not project or not project.subtitles or not project.subtitles.originals:
             raise CommandError(_("No subtitles to batch"), command=self)
 
-        if self.preprocess_subtitles:
-            originals = deepcopy(project.subtitles.originals)
-            preprocessor = SubtitleProcessor(self.options)
-            project.subtitles.PreProcess(preprocessor)
+        with SubtitleEditor(project.subtitles) as editor:
+            if self.preprocess_subtitles:
+                originals = deepcopy(project.subtitles.originals)
+                preprocessor = SubtitleProcessor(self.options)
+                editor.PreProcess(preprocessor)
 
-            if self.options.get('save_preprocessed', False):
-                changed = len(originals) != len(project.subtitles.originals) or any(o != n for o, n in zip(originals, project.subtitles.originals))
-                if changed:
-                    output_path = GetOutputPath(project.subtitles.sourcepath, "preprocessed", project.subtitles.file_format)
-                    logging.info(f"Saving preprocessed subtitles to {output_path}")
-                    project.SaveOriginal(output_path)
+                if self.options.get('save_preprocessed', False):
+                    changed = len(originals) != len(project.subtitles.originals) or any(o != n for o, n in zip(originals, project.subtitles.originals))
+                    if changed:
+                        output_path = GetOutputPath(project.subtitles.sourcepath, "preprocessed", project.subtitles.file_format)
+                        logging.info(f"Saving preprocessed subtitles to {output_path}")
+                        project.SaveOriginal(output_path)
 
-        batcher : SubtitleBatcher = SubtitleBatcher(self.options)
-        project.subtitles.AutoBatch(batcher)
+            batcher : SubtitleBatcher = SubtitleBatcher(self.options)
+            editor.AutoBatch(batcher)
 
         if project.use_project_file:
             self.commands_to_queue.append(SaveProjectFile(project=project))
