+++ PySubtrans/UnitTests/test_Translator.py
@@ -5,6 +5,7 @@ from PySubtrans.Helpers.TestCases import DummyProvider, PrepareSubtitles, Subtit
 from PySubtrans.Helpers.Tests import log_info, log_input_expected_result, log_test_name
 from PySubtrans.SubtitleBatch import SubtitleBatch
 from PySubtrans.SubtitleBatcher import SubtitleBatcher
+from PySubtrans.SubtitleEditor import SubtitleEditor
 from PySubtrans.Subtitles import Subtitles
 from PySubtrans.SubtitleScene import SubtitleScene
 from PySubtrans.SubtitleTranslator import SubtitleTranslator
@@ -33,8 +34,10 @@ class SubtitleTranslatorTests(SubtitleTestCase):
             self.assertEqual(originals.linecount, reference.linecount)
 
             batcher = SubtitleBatcher(self.options)
-            originals.AutoBatch(batcher)
-            reference.AutoBatch(batcher)
+            with SubtitleEditor(originals) as editor:
+                editor.AutoBatch(batcher)
+            with SubtitleEditor(reference) as editor:
+                editor.AutoBatch(batcher)
 
             self.assertEqual(len(originals.scenes), len(reference.scenes))
 
@@ -106,8 +109,10 @@ class SubtitleTranslatorTests(SubtitleTestCase):
             self.assertEqual(originals.linecount, reference.linecount)
 
             batcher = SubtitleBatcher(self.options)
-            originals.AutoBatch(batcher)
-            reference.AutoBatch(batcher)
+            with SubtitleEditor(originals) as editor:
+                editor.AutoBatch(batcher)
+            with SubtitleEditor(reference) as editor:
+                editor.AutoBatch(batcher)
 
             self.assertEqual(len(originals.scenes), len(reference.scenes))
